requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
Appium-Python-Client==1.3.0
attrs==21.4.0
execnet==1.9.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==6.2.5
pytest-forked==1.4.0
pytest-xdist==2.5.0
sauceclient==1.0.0
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.10
